refactor: rename services and remove unused imports.
